extracted compressor func into sepearate file
